Rename git dir client to "file", as the scheme used for it
