+++ formats/file/refs.go
@@ -1,4 +1,4 @@
-package gitdir
+package file
 
 import (
 	"bufio"
